custom TF.keras roi_pooling issue
My fellow kaggler, reported a performance issue of a custom tf pooling module when executed in gpu. Let's check if the model runs in GPU powered colab. I wrote a CIFAR10 classification pipeline with the custom tf pooling module and ran in Colab with GPU. Yes, it ran cool !!!! But the loss results is GARBAGE.